🛠fix: #3 error handling in controller

+++ backend/learnway/src/main/java/com/ssafy/learnway/controller/FriendController.java
@@ -4,6 +4,7 @@ import com.ssafy.learnway.domain.friend.Friend;
 import com.ssafy.learnway.domain.user.User;
 import com.ssafy.learnway.service.FriendService;
 import com.ssafy.learnway.service.UserService;
+import com.ssafy.learnway.util.ResponseHandler;
 import io.swagger.annotations.Api;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.exception.DataException;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.*;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 @Api(tags = {"friend"})
 @RestController
@@ -27,52 +29,68 @@ public class FriendController {
     UserService userService;
 
     @GetMapping("/list")
-    public ResponseEntity list(@RequestBody String userEmail) throws SQLException {
+    public ResponseEntity list(@RequestParam String userEmail) throws SQLException {
         try {
-            Long userId = userService.findByEmail(userEmail).getUserId();
-            List<Friend> friendList = friendService.list(userId);
+            User user = userService.findByEmail(userEmail);
+            Long userId = user.getUserId();
+            List<Friend> friendList = friendService.list(user);
             List<String> userEmailList = new ArrayList<>();
+
 
             // 행 하나의 두 아이디가 검색하려는 유저의 아이디와 다르면 (자신이 아닌 친구라면)
             // 리스트에 저장
             for (Friend friend : friendList) {
                 User tmpUser = friend.getUserId();
                 User tmpFriend = friend.getFriendId();
-                if (tmpUser.getUserId() == userId) userEmailList.add(tmpUser.getUserEmail());
-                else userEmailList.add(tmpFriend.getUserEmail());
+                if (tmpUser.getUserId() == userId) userEmailList.add(tmpFriend.getUserEmail());
+                else userEmailList.add(tmpUser.getUserEmail());
             }
-            return new ResponseEntity<>(userEmailList, HttpStatus.OK);
-        } catch (DataException e) {
+
+            if(userEmailList.isEmpty()) return ResponseHandler.generateResponse("검색된 친구가 없습니다.", HttpStatus.NOT_FOUND);
+
+            return ResponseHandler.generateResponse("검색된 친구 목록입니다.", HttpStatus.OK, "userEmailList", userEmailList);
+        } catch (Exception e) {
             e.printStackTrace();
-            return new ResponseEntity(HttpStatus.NOT_FOUND);
+            return ResponseHandler.generateResponse("서버 오류입니다.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
 
     }
 
     @PostMapping
-    public ResponseEntity make(@RequestBody String user_email, @RequestBody String friend_email) throws SQLException {
+    public ResponseEntity make(String userEmail, String friendEmail) throws SQLException {
         try {
             // 유저, 친구 객체 반환
-            User user = userService.findByEmail(user_email);
-            User friend = userService.findByEmail(friend_email);
+            User user = userService.findByEmail(userEmail);
+            User friend = userService.findByEmail(friendEmail);
+
+            if (user==null || friend == null) return ResponseHandler.generateResponse("유효하지 않은 이메일입니다.", HttpStatus.NOT_FOUND);
+
             Friend newRelation = Friend.builder().userId(user).friendId(friend).build();
             friendService.make(newRelation);
-        } catch (DataException e) {
+
+            return ResponseHandler.generateResponse("친구 생성에 성공하였습니다.", HttpStatus.OK);
+        } catch (Exception e) {
             e.printStackTrace();
-            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+            return ResponseHandler.generateResponse("친구 생성에 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
-        return new ResponseEntity(HttpStatus.OK);
+
     }
 
     @DeleteMapping
-    public void delete(@RequestBody String user_email, String friend_email) throws SQLException {
+    public ResponseEntity delete(String user_email, String friend_email){
         try {
-            Long userId = userService.findByEmail(user_email).getUserId();
-            Long friendId = userService.findByEmail(friend_email).getUserId();
-            Friend relation = friendService.findByEmail(userId, friendId);
+            User user = userService.findByEmail(user_email);
+            User friend = userService.findByEmail(friend_email);
+
+            if (user==null || friend == null) return ResponseHandler.generateResponse("유효하지 않은 이메일입니다.", HttpStatus.NOT_FOUND);
+
+            Friend relation = friendService.findById(user, friend);
+            System.out.println(relation.toString());
             friendService.delete(relation);
-        } catch (DataException e) {
+            return ResponseHandler.generateResponse("친구가 삭제되었습니다.", HttpStatus.OK);
+        } catch (Exception e) {
             e.printStackTrace();
+            return ResponseHandler.generateResponse("친구 삭제에 실패하였습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 }
